refactor: optimize session context, split file_tools module, and fix harbor paused hang
