+++ src-tauri/src/mcp/builtin/workspace/tools/file_tools/mod.rs
@@ -1,0 +1,7 @@
+pub mod edit;
+pub mod read;
+pub mod write;
+
+pub use edit::*;
+pub use read::*;
+pub use write::*;
